Test: update 'The Menu' suite spec to be more robust
Initial spec, 'visibility is toggled on click', assumed spec 'is hidden by default' passed. E.g. originally, the former spec tested if the hidden state of the menu toggled from true to false, back to true, on click. This assumed that the state started initially as true. The new spec only tests that the state toggles on click, independent of initial state value.