Test: restore 'The Menu' spec to more strict form
This is the form I had it initially. I had changed it to be more robust, so that it would test only the toggling of the menu's visibility, regardless of the menu's initial visibility. However, project requirements prefer that the spec test specifically the toggling of menu visibility *based on* initial visibility state.