Test: remove redundant expects() & simplify other expects()
I realize from experimentation that Jasmine will catch errors if a variable used in a spec is undefined, so I don't need to check for each variables definition in the 'New Feed Selection' and 'Initial Entries' suites.

I also don't need to check for a variable to be defined and also check its length with two seperate expects. I can just check if said variable is truthy. Made some minor simplifications like this.